requirements.txt
torch
torchvision
hydra-core
omegaconf
wandb
h5py
pytest
ultralytics
